DRChanllenge/PictureDetect/requirements.txt
Pillow==8.2.0
numpy==1.20.2
opencv-python==4.5.1.48